Corrige alertas
